+++ package.json
@@ -13,7 +13,7 @@
     "mysql": "^2.15.0"
   },
   "devDependencies": {
-    "angular": "^1.6.6",
+    "angular": "^1.6.7",
     "angular-ui-router": "^0.4.3",
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.0.2",
@@ -24,19 +24,21 @@
     "bootstrap-sass": "^3.3.7",
     "clean-webpack-plugin": "^0.1.17",
     "css-loader": "^0.28.7",
-    "express": "^4.15.5",
+    "eslint-loader": "^1.9.0",
+    "express": "^4.16.2",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.5",
     "html-loader": "^0.5.1",
     "imports-loader": "^0.7.1",
+    "jquery": "^3.2.1",
     "node-sass": "^4.6.1",
-    "postcss-loader": "^2.0.8",
+    "postcss-loader": "^2.0.9",
     "protractor": "^5.2.0",
     "resolve-url-loader": "^2.2.0",
     "sass-loader": "^6.0.6",
     "standard-loader": "^6.0.1",
     "style-loader": "^0.19.0",
-    "uglifyjs-webpack-plugin": "^1.0.1",
+    "uglifyjs-webpack-plugin": "^1.1.1",
     "url-loader": "^0.6.2",
     "webpack": "^3.8.1",
     "webpack-dev-server": "^2.9.4"
